Conflicts:
	nts.uk/uk.at/at.web/nts.uk.at.web/src/main/webapp/WEB-INF/web.xml
	nts.uk/uk.base/base.web/nts.uk.com.js.web/src/main/webapp/lib/nittsu/nts.uk.com.web.nittsu.bundles.js.map
	nts.uk/uk.com/com.ctx/sys/auth/nts.uk.ctx.sys.auth.pubimp/src/main/java/nts/uk/ctx/sys/auth/pubimp/grant/RoleSetGrantedEmployeePubImpl.java
	nts.uk/uk.com/com.web/nts.uk.com.web/src/main/webapp/WEB-INF/web.xml
	nts.uk/uk.repo/nts.uk.com.web.jar
	nts.uk/uk.repo/nts.uk.ctx.basic.ac.jar
	nts.uk/uk.repo/nts.uk.ctx.basic.app.jar
	nts.uk/uk.repo/nts.uk.ctx.basic.dom.jar
	nts.uk/uk.repo/nts.uk.ctx.basic.infra.jar
	nts.uk/uk.repo/nts.uk.ctx.basic.pub.jar
	nts.uk/uk.repo/nts.uk.ctx.basic.pubimp.jar
	nts.uk/uk.repo/nts.uk.ctx.basic.ws.jar
	nts.uk/uk.repo/nts.uk.ctx.bs.company.ac.jar
	nts.uk/uk.repo/nts.uk.ctx.bs.company.app.jar
	nts.uk/uk.repo/nts.uk.ctx.bs.company.dom.jar
	nts.uk/uk.repo/nts.uk.ctx.bs.company.infra.jar
	nts.uk/uk.repo/nts.uk.ctx.bs.company.pub.jar
	nts.uk/uk.repo/nts.uk.ctx.bs.company.pubimp.jar
	nts.uk/uk.repo/nts.uk.ctx.bs.company.ws.jar
	nts.uk/uk.repo/nts.uk.ctx.bs.employee.ac.jar
	nts.uk/uk.repo/nts.uk.ctx.bs.employee.app.jar
	nts.uk/uk.repo/nts.uk.ctx.bs.employee.dom.jar
	nts.uk/uk.repo/nts.uk.ctx.bs.employee.infra.jar
	nts.uk/uk.repo/nts.uk.ctx.bs.employee.pub.jar
	nts.uk/uk.repo/nts.uk.ctx.bs.employee.pubimp.jar
	nts.uk/uk.repo/nts.uk.ctx.bs.employee.ws.jar
	nts.uk/uk.repo/nts.uk.ctx.bs.person.ac.jar
	nts.uk/uk.repo/nts.uk.ctx.bs.person.app.jar
	nts.uk/uk.repo/nts.uk.ctx.bs.person.dom.jar
	nts.uk/uk.repo/nts.uk.ctx.bs.person.infra.jar
	nts.uk/uk.repo/nts.uk.ctx.bs.person.pub.jar
	nts.uk/uk.repo/nts.uk.ctx.bs.person.pubimp.jar
	nts.uk/uk.repo/nts.uk.ctx.bs.person.ws.jar
	nts.uk/uk.repo/nts.uk.ctx.exio.ac.jar
	nts.uk/uk.repo/nts.uk.ctx.exio.app.jar
	nts.uk/uk.repo/nts.uk.ctx.exio.dom.jar
	nts.uk/uk.repo/nts.uk.ctx.exio.infra.jar
	nts.uk/uk.repo/nts.uk.ctx.exio.ws.jar
	nts.uk/uk.repo/nts.uk.ctx.pereg.ac.jar
	nts.uk/uk.repo/nts.uk.ctx.pereg.app.jar
	nts.uk/uk.repo/nts.uk.ctx.pereg.dom.jar
	nts.uk/uk.repo/nts.uk.ctx.pereg.infra.jar
	nts.uk/uk.repo/nts.uk.ctx.pereg.ws.jar
	nts.uk/uk.repo/nts.uk.ctx.sys.auth.ac.jar
	nts.uk/uk.repo/nts.uk.ctx.sys.auth.app.jar
	nts.uk/uk.repo/nts.uk.ctx.sys.auth.dom.jar
	nts.uk/uk.repo/nts.uk.ctx.sys.auth.infra.jar
	nts.uk/uk.repo/nts.uk.ctx.sys.auth.pub.jar
	nts.uk/uk.repo/nts.uk.ctx.sys.auth.pubimp.jar
	nts.uk/uk.repo/nts.uk.ctx.sys.auth.ws.jar
	nts.uk/uk.repo/nts.uk.ctx.sys.env.ac.jar
	nts.uk/uk.repo/nts.uk.ctx.sys.env.app.jar
	nts.uk/uk.repo/nts.uk.ctx.sys.env.dom.jar
	nts.uk/uk.repo/nts.uk.ctx.sys.env.infra.jar
	nts.uk/uk.repo/nts.uk.ctx.sys.env.pub.jar
	nts.uk/uk.repo/nts.uk.ctx.sys.env.pubimp.jar
	nts.uk/uk.repo/nts.uk.ctx.sys.env.ws.jar
	nts.uk/uk.repo/nts.uk.ctx.sys.gateway.ac.jar
	nts.uk/uk.repo/nts.uk.ctx.sys.gateway.app.jar
	nts.uk/uk.repo/nts.uk.ctx.sys.gateway.dom.jar
	nts.uk/uk.repo/nts.uk.ctx.sys.gateway.infra.jar
	nts.uk/uk.repo/nts.uk.ctx.sys.gateway.pub.jar
	nts.uk/uk.repo/nts.uk.ctx.sys.gateway.pubimp.jar
	nts.uk/uk.repo/nts.uk.ctx.sys.gateway.ws.jar
	nts.uk/uk.repo/nts.uk.ctx.sys.portal.ac.jar
	nts.uk/uk.repo/nts.uk.ctx.sys.portal.app.jar
	nts.uk/uk.repo/nts.uk.ctx.sys.portal.dom.jar
	nts.uk/uk.repo/nts.uk.ctx.sys.portal.infra.jar
	nts.uk/uk.repo/nts.uk.ctx.sys.portal.pub.jar
	nts.uk/uk.repo/nts.uk.ctx.sys.portal.pubimp.jar
	nts.uk/uk.repo/nts.uk.ctx.sys.portal.ws.jar
	nts.uk/uk.repo/nts.uk.ctx.sys.shared.app.jar
	nts.uk/uk.repo/nts.uk.ctx.sys.shared.dom.jar
	nts.uk/uk.repo/nts.uk.ctx.sys.shared.infra.jar
	nts.uk/uk.repo/nts.uk.ctx.sys.shared.ws.jar
	nts.uk/uk.repo/nts.uk.ctx.workflow.ac.jar
	nts.uk/uk.repo/nts.uk.ctx.workflow.app.jar
	nts.uk/uk.repo/nts.uk.ctx.workflow.dom.jar
	nts.uk/uk.repo/nts.uk.ctx.workflow.infra.jar
	nts.uk/uk.repo/nts.uk.ctx.workflow.pub.jar
	nts.uk/uk.repo/nts.uk.ctx.workflow.pubimp.jar
	nts.uk/uk.repo/nts.uk.ctx.workflow.ws.jar
	nts.uk/uk.repo/nts.uk.file.com.app.jar
	nts.uk/uk.repo/nts.uk.file.com.infra.jar
	nts.uk/uk.repo/nts.uk.file.com.ws.jar
	nts.uk/uk.repo/nts.uk.query.ac.jar
	nts.uk/uk.repo/nts.uk.query.app.jar
	nts.uk/uk.repo/nts.uk.query.infra.jar
	nts.uk/uk.repo/nts.uk.query.model.jar
	nts.uk/uk.repo/nts.uk.query.pub.jar
	nts.uk/uk.repo/nts.uk.query.pubimp.jar
	nts.uk/uk.repo/nts.uk.query.ws.jar
	nts.uk/uk.repo/nts.uk.screen.com.app.jar
	nts.uk/uk.repo/nts.uk.screen.com.infra.jar
	nts.uk/uk.repo/nts.uk.screen.com.ws.jar
	nts.uk/uk.repo/nts.uk.shr.com.jar
	nts.uk/uk.repo/nts.uk.shr.find.jar
	nts.uk/uk.repo/nts.uk.shr.ws.jar

+++ nts.uk/uk.at/at.ctx/record/nts.uk.ctx.at.record.infra/src/main/java/nts/uk/ctx/at/record/infra/entity/workrecord/erroralarm/otkcustomize/KrcctOtkVacationCk.java
@@ -33,7 +33,7 @@ public class KrcctOtkVacationCk extends UkJpaEntity implements Serializable {
 	private String cid;
 
 	@Column(name="USE_ATR")
-	public long useAtr;
+	public int useAtr;
 
 	@Column(name="CONTINUOUS_DAYS")
 	public int continuousDays;
@@ -50,7 +50,7 @@ public class KrcctOtkVacationCk extends UkJpaEntity implements Serializable {
 	public KrcctOtkVacationCk() {
 	}
 
-	public KrcctOtkVacationCk(String cid, long useAtr, int continuousDays, String messageDisplay) {
+	public KrcctOtkVacationCk(String cid, int useAtr, int continuousDays, String messageDisplay) {
 		super();
 		this.cid = cid;
 		this.useAtr = useAtr;
